refactor(task-213): moving valueObjects folder to domain level
