+++ backend/src/domain/valueObjects/student/difficulties.ts
